Check disk usage using FileStore API
The previous way of checking gave the "apparent" usage, rather than the
"actual" usage. Since Kafka uses sparse files, this was not an optimal
way of measuring disk usage for our purposes here -- it could report
10s of gigabytes of usage when the reality of the disk usage might
only be a few megabytes.

The main difference between this and shelling out to `du` is that this
reports the space used on the entire disk, rather than just the
directories specified by logDirs. There are advantages and
disadvantages to this.